requirements.txt
django-admin-interface
django-ckeditor
edx-organizations
# mysqlclient a través de pip install mysqlcient
# mysqlclient a través de pip install djangorestframework
# pip install django-phonenumber-field[phonenumbers]